Update: add version

+++ web/src/main/java/com/epam/esm/controller/TagController.java
@@ -7,11 +7,13 @@ import com.epam.esm.exception.EntityNotFoundException;
 import com.epam.esm.exception.InvalidEntityException;
 import com.epam.esm.service.impl.TagServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.i18n.LocaleContextHolder;
+import org.springframework.context.support.ResourceBundleMessageSource;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
+import java.util.*;
 
 
 /**
@@ -24,6 +26,12 @@ import java.util.List;
 public class TagController {
 
     private final TagServiceImpl tagService;
+    private static final String VERSION = "v1.0";
+    private static final String ERROR_MESSAGE = "errorMessage";
+    private static final String ERROR_CODE = "errorCode";
+    private static final String VERSION_ERROR = "versionError";
+
+    private final ResourceBundleMessageSource bundleMessageSource;
 
     /**
      * Instantiates a new Tag controller.
@@ -31,7 +39,8 @@ public class TagController {
      * @param tagService the tag service
      */
     @Autowired
-    public TagController(TagServiceImpl tagService) {
+    public TagController(TagServiceImpl tagService, ResourceBundleMessageSource messageSource) {
+        this.bundleMessageSource = messageSource;
         this.tagService = tagService;
     }
 
@@ -41,9 +50,14 @@ public class TagController {
      * @return JSON {@link ResponseEntity} object that contains list of {@link TagDto}
      */
     @GetMapping
-    public ResponseEntity<List<TagDto>> getAllTags() {
-        List<TagDto> tagDtoList = tagService.findAllTags();
-        return new ResponseEntity<>(tagDtoList, HttpStatus.OK);
+    public ResponseEntity<Object> getAllTags(@PathVariable String ver) {
+        if (Objects.equals(ver, VERSION)){
+            List<TagDto> tagDtoList = tagService.findAllTags();
+            return new ResponseEntity<>(tagDtoList, HttpStatus.OK);
+        }
+        String errorMessage = getErrorMessage(VERSION_ERROR);
+        return buildResponse(HttpStatus.MOVED_PERMANENTLY, errorMessage, HttpStatus.MOVED_PERMANENTLY.toString());
+
     }
 
     /**
@@ -54,9 +68,14 @@ public class TagController {
      * @throws EntityNotFoundException in case certificate with this id doesn't exist
      */
     @GetMapping("/{id}")
-    public ResponseEntity<TagDto> getTag(@PathVariable("id") Long id) {
-        TagDto tagById = tagService.findTagById(id);
-        return new ResponseEntity<>(tagById, HttpStatus.OK);
+    public ResponseEntity<Object> getTag(@PathVariable("id") Long id, @PathVariable String ver) {
+        if (Objects.equals(ver, VERSION)) {
+            TagDto tagById = tagService.findTagById(id);
+            return new ResponseEntity<>(tagById, HttpStatus.OK);
+        }
+        String errorMessage = getErrorMessage(VERSION_ERROR);
+        return buildResponse(HttpStatus.MOVED_PERMANENTLY, errorMessage, HttpStatus.MOVED_PERMANENTLY.toString());
+
     }
 
     /**
@@ -68,9 +87,13 @@ public class TagController {
      * @throws EntityAlreadyExistsException when {@link Tag} entity is already exists
      */
     @PostMapping
-    public ResponseEntity<Void> createTag(@RequestBody TagDto tagDto) {
-        tagService.create(tagDto);
-        return new ResponseEntity<>(HttpStatus.CREATED);
+    public ResponseEntity<Object> createTag(@RequestBody TagDto tagDto, @PathVariable String ver) {
+        if (Objects.equals(ver, VERSION)) {
+            tagService.create(tagDto);
+            return new ResponseEntity<>(HttpStatus.CREATED);
+        }
+        String errorMessage = getErrorMessage(VERSION_ERROR);
+        return buildResponse(HttpStatus.MOVED_PERMANENTLY, errorMessage, HttpStatus.MOVED_PERMANENTLY.toString());
     }
 
     /**
@@ -82,9 +105,25 @@ public class TagController {
      * @throws InvalidEntityException when id is not written correctly
      */
     @DeleteMapping("/{id}")
-    public ResponseEntity<Boolean> deleteTag(@PathVariable("id") Long id) {
-        boolean hasBeenDeleted = tagService.delete(id);
-        return hasBeenDeleted ? new ResponseEntity<>(hasBeenDeleted, HttpStatus.OK)
-                : new ResponseEntity<>(hasBeenDeleted, HttpStatus.NOT_MODIFIED);
+    public ResponseEntity<Object> deleteTag(@PathVariable("id") Long id, @PathVariable String ver) {
+        if (Objects.equals(ver, VERSION)) {
+            boolean hasBeenDeleted = tagService.delete(id);
+            return hasBeenDeleted ? new ResponseEntity<>(hasBeenDeleted, HttpStatus.OK)
+                    : new ResponseEntity<>(hasBeenDeleted, HttpStatus.NOT_MODIFIED);
+        }
+        String errorMessage = getErrorMessage(VERSION_ERROR);
+        return buildResponse(HttpStatus.MOVED_PERMANENTLY, errorMessage, HttpStatus.MOVED_PERMANENTLY.toString());
+    }
+
+    private String getErrorMessage(String errorMessageName) {
+        Locale locale = LocaleContextHolder.getLocale();
+        return bundleMessageSource.getMessage(errorMessageName, null, locale);
+    }
+
+    private ResponseEntity<Object> buildResponse(HttpStatus status, String errorMessage, String errorCode) {
+        Map<String, Object> map = new HashMap<>();
+        map.put(ERROR_MESSAGE, errorMessage);
+        map.put(ERROR_CODE, errorCode);
+        return new ResponseEntity<>(map, status);
     }
 }
